Simplify code to obtain the position of all the planets
Instead of storing the positions of all the plants in a DynamicBuffer,
I use a Tag Component for the plannets. On the SpawningSystem, I look
for the closest planet for each spaceship, and store its position
instead of storing the planet waypoint. This reduce the ammount of code,
and makes the code more efficient.